Default to 'main' branch instead of 'master'
As described in https://github.com/github/renaming, the default branch
on github is moving from master to main and tooling will be available
later this year to transition existing repos. So more people will be
able to use the defaults if we adjust the defaults here.

Also, mark the branch as no longer required with these sane defaults.